updates and fcpu macro

+++ ntiny_ide/ntiny_application/bsp/src/clock.c
@@ -4,7 +4,7 @@
 
 clock_t clock()
 {
-	unsigned long long cycles = csr_read(mcycleh);
+	unsigned long cycles = csr_read(mcycleh);
   	cycles = (cycles << 32) + csr_read(mcycle);
   	return (clock_t)cycles;
 }
